+++ lib/influxdb2/client/client.rb
@@ -45,6 +45,7 @@ module InfluxDB2
     # @option options [bool] :use_ssl Turn on/off SSL for HTTP communication
     #   the body line-protocol
     def initialize(url, token, options = nil)
+      @auto_closeable = []
       @options = options ? options.dup : {}
       @options[:url] = url if url.is_a? String
       @options[:token] = token if token.is_a? String
@@ -56,8 +57,10 @@ module InfluxDB2
     # Write time series data into InfluxDB thought WriteApi.
     #
     # @return [WriteApi] New instance of WriteApi.
-    def create_write_api
-      WriteApi.new(options: @options)
+    def create_write_api(write_options: InfluxDB2::SYNCHRONOUS)
+      write_api = WriteApi.new(options: @options, write_options: write_options)
+      @auto_closeable.push(write_api)
+      write_api
     end
 
     # Get the Query client.
@@ -72,6 +75,7 @@ module InfluxDB2
     # @return [ true ] Always true.
     def close!
       @closed = true
+      @auto_closeable.each(&:close!)
       true
     end
   end