+++ lib/influxdb2/client/write_api.rb
@@ -17,8 +17,31 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
+require_relative 'worker'
 
 module InfluxDB2
+  module WriteType
+    SYNCHRONOUS = 1
+    BATCHING = 2
+  end
+
+  # Creates write api configuration.
+  #
+  # @param write_type: methods of write (batching, asynchronous, synchronous)
+  # @param batch_size: the number of data point to collect in batch
+  # @param flush_interval: flush data at least in this interval
+  class WriteOptions
+    def initialize(write_type: WriteType::SYNCHRONOUS, batch_size: 1_000, flush_interval: 1_000)
+      @write_type = write_type
+      @batch_size = batch_size
+      @flush_interval = flush_interval
+    end
+
+    attr_reader :write_type, :batch_size, :flush_interval
+  end
+
+  SYNCHRONOUS = InfluxDB2::WriteOptions.new(write_type: WriteType::SYNCHRONOUS)
+
   # Precision constants.
   #
   class WritePrecision
@@ -39,9 +62,13 @@ module InfluxDB2
   #
   class WriteApi < DefaultApi
     # @param [Hash] options The options to be used by the client.
-    def initialize(options:)
+    # @param [WriteOptions] write_options Write api configuration.
+    def initialize(options:, write_options: SYNCHRONOUS)
       super(options: options)
+      @write_options = write_options
+      @closed = false
     end
+    attr_reader :closed
 
     # Write data into specified Bucket.
     #
@@ -83,8 +110,36 @@ module InfluxDB2
       _check('bucket', bucket_param)
       _check('org', org_param)
 
-      payload = _generate_payload(data)
+      payload = _generate_payload(data, bucket: bucket_param, org: org_param, precision: precision_param)
       return nil if payload.nil?
+
+      if WriteType::BATCHING == @write_options.write_type
+        _worker.push(payload)
+      else
+        write_raw(payload, precision: precision_param, bucket: bucket_param, org: org_param)
+      end
+    end
+
+    # @return [ true ] Always true.
+    def close!
+      _worker.flush_all unless _worker.nil?
+      @closed = true
+      true
+    end
+
+    # @param [String] payload data as String
+    # @param [WritePrecision] precision The precision for the unix timestamps within the body line-protocol
+    # @param [String] bucket specifies the destination bucket for writes
+    # @param [String] org specifies the destination organization for writes
+    def write_raw(payload, precision: nil, bucket: nil, org: nil)
+      precision_param = precision || @options[:precision]
+      bucket_param = bucket || @options[:bucket]
+      org_param = org || @options[:org]
+      _check('precision', precision_param)
+      _check('bucket', bucket_param)
+      _check('org', org_param)
+
+      return nil unless payload.instance_of?(String) || payload.empty?
 
       uri = URI.parse(File.join(@options[:url], '/api/v2/write'))
       uri.query = URI.encode_www_form(bucket: bucket_param, org: org_param, precision: precision_param.to_s)
@@ -92,24 +147,71 @@ module InfluxDB2
       _post(payload, uri)
     end
 
+    # Item for batching queue
+    class BatchItem
+      def initialize(key, data)
+        @key = key
+        @data = data
+      end
+      attr_reader :key, :data
+    end
+
+    # Key for batch item
+    class BatchItemKey
+      def initialize(bucket, org, precision = DEFAULT_WRITE_PRECISION)
+        @bucket = bucket
+        @org = org
+        @precision = precision
+      end
+      attr_reader :bucket, :org, :precision
+
+      def ==(other)
+        @bucket == other.bucket && @org == other.org && @precision == other.precision
+      end
+
+      alias eql? ==
+
+      def hash
+        @bucket.hash ^ @org.hash ^ @precision.hash # XOR
+      end
+    end
+
     private
 
-    def _generate_payload(data)
+    WORKER_MUTEX = Mutex.new
+    def _worker
+      return nil unless @write_options.write_type == WriteType::BATCHING
+
+      return @worker if @worker
+
+      WORKER_MUTEX.synchronize do
+        # this return is necessary because the previous mutex holder
+        # might have already assigned the @worker
+        return @worker if @worker
+
+        @worker = Worker.new(self, @write_options)
+      end
+    end
+
+    def _generate_payload(data, precision: nil, bucket: nil, org: nil)
       if data.nil?
         nil
       elsif data.is_a?(Point)
-        data.to_line_protocol
+        _generate_payload(data.to_line_protocol, bucket: bucket, org: org, precision: data.precision ||
+            DEFAULT_WRITE_PRECISION)
       elsif data.is_a?(String)
         if data.empty?
           nil
+        elsif @write_options.write_type == WriteType::BATCHING
+          BatchItem.new(BatchItemKey.new(bucket, org, precision), data)
         else
           data
         end
       elsif data.is_a?(Hash)
-        _generate_payload(Point.from_hash(data))
+        _generate_payload(Point.from_hash(data), bucket: bucket, org: org, precision: precision)
       elsif data.respond_to? :map
         data.map do |item|
-          _generate_payload(item)
+          _generate_payload(item, bucket: bucket, org: org, precision: precision)
         end.reject(&:nil?).join("\n".freeze)
       end
     end